resume-rag-lab/requirements.txt
streamlit
langchain
langchain-openai
faiss-cpu
pypdf
tiktoken
python-dotenv